Correct flaky distributed test. (#56)
The test was intermittently failing because the transaction in the data cube may not have had enough time to be received and processed by the query cubes.
As a consequence, there were no chunks in the Query cubes, hence no entries for them in the chunk table.

Although using a standard method like the `QueryCubeSync` would be better to synchronize Data and Query cubes, it was not easy to make it work. A basic wait-and-see proved far more eefficient. Thus, we retain this solution.