Ignore conflicting slicing hierarchies in smart filtering.
The combination of the two hierarchies was creating cases where no data
where display, because the filters where creating empty results.
In addition, we order the Branch hierarchy to have master as the first
value, picked to be the default member. This is a sensible choice
because every dump will have it as a branch.